requirements.txt
flask==3.0.0
torch>=2.1.2
transformers>=4.36.2
accelerate>=0.25.0
python-dotenv==1.0.0
gunicorn==21.2.0